requirements.txt
drf-yasg==1.21.7
djangorestframework==3.14.0
Django==3.2.12